Corrección Modelo de datos - Valoraciones
* Se corrigen algunos aspectos de la tabla: campo valoración - se agregan las claves.
* Se corrige el codigo para que sea consistente con las modificaciones en dicha tabla.

+++ TeLoBusco/Servicios/AccesoDatos/ValoracionesServicios.cs
@@ -37,7 +37,7 @@ namespace Servicios.AccesoDatos
 
             }
         }
-        public static decimal ObtenerPuntaje(string id)
+        public static double ObtenerPuntaje(string id)
         {
             using (TeloBuscoEntities db = new TeloBuscoEntities())
             {
